backend/requirements.txt
fastapi
uvicorn[standard]
requests
python-dotenv
httpx
pydantic
sentence-transformers
faiss-cpu
sqlalchemy
